% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_branch.R
\name{actualizar_rama}
\alias{actualizar_rama}
\title{Actualice la rama master desde el upstream}
\usage{
actualizar_rama()
}
\value{
No hay valor de retorno - se llama por su efecto secundario
}
\description{
actualiza su rama master con su upstream
por medio de git rebase
}
