% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{add_token}
\alias{add_token}
\title{Autentificación en ixplorer}
\usage{
add_token()
}
\description{
Hace la conexión a su repositorio a través del gadget de ixplorer y
permite crear y revisar tiquetes sin tener que re-ingresar sus credenciales
}
