% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{add_upstream}
\alias{add_upstream}
\title{Add the upstream to the repository that is active in RStudio.}
\usage{
add_upstream(instance = "saved")
}
\arguments{
\item{instance}{ixplorer instance (Eg: "secure", "masterclass", "prueba")}
}
\description{
Add the upstream to the repository that is active in RStudio at
the moment
}
