% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_users.R
\name{add_user_to_repo}
\alias{add_user_to_repo}
\title{Add a user as a collaborator to an ixplorer repository.}
\usage{
add_user_to_repo(
  base_url,
  api_key,
  owner,
  repo,
  username,
  collaborator_permissions
)
}
\arguments{
\item{base_url}{The base URL of the ixplorer server.}

\item{api_key}{The API key for authentication on the ixplorer server.}

\item{owner}{The owner (organization or user) of the repository where
the user will be added as a collaborator.}

\item{repo}{The name of the repository where the user will be added
as a collaborator.}

\item{username}{The username of the user to be added as a collaborator.}

\item{collaborator_permissions}{The permissions to be assigned
to the collaborator ("Read", "Write", or "Admin").}
}
\value{
The response from the ixplorer server.
}
\description{
This function adds a user as a collaborator to a specific repository on
ixplorer with specific permissions.
}
