% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ticket.R
\name{crear_tiquetes}
\alias{crear_tiquetes}
\title{Crear tiquete}
\usage{
crear_tiquetes(instancia = "guardada")
}
\arguments{
\item{instancia}{instancia de ixplorer (Ejemplo: "secure", "masterclass",
"prueba")}
}
\value{
No hay valor de retorno - se llama por su efecto secundario
}
\description{
Cree tiquetes (título y cuerpo) desde el add-in de ixplorer sin
perder las ideas durante su flujo de trabajo
}
