% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_issue.R
\name{create_issue}
\alias{create_issue}
\title{Create issue}
\usage{
create_issue()
}
\description{
Create issues (Title and body) from the ixplorer addin without loosing the
ideas during your workflow. Issues will be in the repository that corresponds
to the information youo give in the authenticate gadget.
}
