% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ticket.R
\name{create_tickets}
\alias{create_tickets}
\title{Crear tiquete}
\usage{
create_tickets()
}
\description{
Crear tiquetes (Título y cuerpo) desde el addin de ixplorer sin perder las
ideas durante su flujo de trabajo. Los tiquetes serán creados en el
repositorio que corresponde a la información dada en el gadget de
autentificación.
}
