% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_credentials.R
\name{delete_credentials}
\alias{delete_credentials}
\title{Delete your ixplorer credentials}
\usage{
delete_credentials(instance = "current")
}
\arguments{
\item{instance}{an instance from ixplorer such as "prueba" or "masterclass"}
}
\value{
Delete credentials saved by add_token()
}
\description{
Delete your ixplorer credentials
}
\details{
Delete the credentials saved by add_token() in your computer.
This can be useful when you are using the ixplorer package on a
shared computer or any other case you don't want to keep the credentials
stored
}
\examples{
\dontrun{
delete_credentials("masterclass")
}
}
