% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{fijar_pausa_git}
\alias{fijar_pausa_git}
\title{Extienda el tiempo de sus credenciales de git}
\usage{
fijar_pausa_git(pausa = 14000, global = FALSE)
}
\arguments{
\item{global}{}
}
\description{
Esta es una traducida que llama a set_git_timeout().
Establezca la cantidad de segundos para que se agote el
tiempo de espera del caché de credenciales de git. Tenga en cuenta que
esto está escrito para trabajar en servidores remotos donde no queremos
almacenar nuestras credenciales de git (otros pueden tener acceso o
hacerse pasar por nosotros). En una máquina confiable, la configuración más
común sería (en una máquina Linux) <git credential.helper store>. Sin
embargo, se requiere algo de almacenamiento en caché cuando se trabaja en
un servidor remoto porque los ciclos frecuentes de envío y confirmación en
los que tiene que escribir su nombre de usuario y contraseña cada vez no son
agradables.

De forma predeterminada, esto solo se aplicará al repositorio en el que está
trabajando. Si llama a la versión global, asegúrese de que esto es lo que
desea. También por defecto, el tiempo de espera se establece en 4 horas
(14400 segundos).
}
