% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instance.R
\name{get_instance}
\alias{get_instance}
\title{Get the active instance in the session}
\usage{
get_instance()
}
\value{
a character of the active issue in the session
}
\description{
By default, functions that get tickets and pull requests
use the active instance in the session. This functions gets this
active instance.
}
