% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{incluye_upstream}
\alias{incluye_upstream}
\title{A침ade el upstream al repositorio que est치 activo en RStudio}
\usage{
incluye_upstream(instance, owner)
}
\arguments{
\item{instance}{instancia de ixplorer (Ejm: "secure", "masterclass", "prueba")}

\item{owner}{el nombre del proyecto donde se encuentra el repositorio en
ixplorer}
}
\description{
A침ade el upstream al repositorio que est치 activo en RStudio en
este momento
}
