% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixplorer.R
\name{ix_reports}
\alias{ix_reports}
\title{ixplorer reports}
\usage{
ix_reports()
}
\description{
Visualize issues of an specific user, a team and closed issues based on
the credentials used in gadget authenticate.
}
