% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixplorer.R
\name{ix_tickets}
\alias{ix_tickets}
\title{ixplorer reports}
\usage{
ix_tickets()
}
\description{
Visualize the tickets of an specific user, a team and get the quick links to
your ixplorer based on the credentials used in gadget authenticate.
}
