% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixplorer-package.R
\docType{package}
\name{ixplorer-package}
\alias{ixplorer}
\alias{ixplorer-package}
\title{ixplorer: Easy DataOps for R Users}
\description{
Create and view tickets in 'gitea', a self-hosted git service \url{https://gitea.io}, using an 'RStudio' addin, and use helper functions to publish documentation and use git.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ixpantia/ixplorer}
  \item Report bugs at \url{https://github.com/ixpantia/ixplorer/issues}
}

}
\author{
\strong{Maintainer}: Frans van Dunne \email{frans@ixpantia.com} (\href{https://orcid.org/0000-0002-7853-2811}{ORCID})

Authors:
\itemize{
  \item Ronny Hernández Mora \email{ronny@ixpantia.com} (\href{https://orcid.org/0000-0001-6225-7096}{ORCID})
}

Other contributors:
\itemize{
  \item ixpantia, SRL \email{hola@ixpantia.com} [copyright holder]
  \item Daniel Granados Campos \email{daniel@ixpantia.com} [contributor]
  \item Nayib Vargas Zuñiga \email{nayib@ixpantia.com} [contributor]
}

}
\keyword{internal}
