% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{list_closed_tickets}
\alias{list_closed_tickets}
\title{Lista de tiquetes cerrados}
\usage{
list_closed_tickets(instance, owner, repository = "current", lag = 7)
}
\arguments{
\item{instance}{instancia de ixplorer (Ejm: "secure", "masterclass", "prueba")}

\item{owner}{el nombre del proyecto donde se encuentra el repositorio en
ixplorer}

\item{repository}{el nombre del repositorio donde están los tiquetes}
}
\description{
Listado de tiquetes cerrados del repositorio indicado.
}
