% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{list_tickets}
\alias{list_tickets}
\alias{list_open_tickets}
\alias{list_closed_tickets}
\title{List open and closed tickets.}
\usage{
list_open_tickets(lag = 7, repository = "current", clip = TRUE)

list_closed_tickets(lag = 7, repository = "current", clip = TRUE)
}
\arguments{
\item{lag}{FIXME}

\item{repository}{FIXME}

\item{clip}{FIXME}
}
\value{
FIXME
}
\description{
\code{list_open_tickets()} and \code{list_closed_tickets} list open and closed
tickets. The "current" respository is the one linked to the active RStudio
project.
}
