% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_open_tickets.R
\name{listar_tiquetes_abiertos}
\alias{listar_tiquetes_abiertos}
\title{Lista de tiquetes abiertos}
\usage{
listar_tiquetes_abiertos(instancia = "guardada", dias = NULL)
}
\arguments{
\item{instance}{instancia de ixplorer (Ejm: "secure", "masterclass", "prueba")}

\item{owner}{el nombre del proyecto donde se encuentra el repositorio en
ixplorer}

\item{repository}{el nombre del repositorio donde están los tiquetes}

\item{lag}{es el tiempo en días que se quiere ver hacia atrás. Por ejemplo si
se quiere ver los tiquetes creados en los últimos 7 días, lag = 7. Por
defecto muestra todos los tiquetes sin ningún lag.}
}
\description{
Listado de tiquetes abiertos del repositorio indicado.
}
