% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_closed_tickets.R
\name{listar_tiquetes_cerrados}
\alias{listar_tiquetes_cerrados}
\title{Lista de tiquetes cerrados}
\usage{
listar_tiquetes_cerrados(instancia = "guardada", dias = NULL)
}
\arguments{
\item{instancia}{instancia de ixplorer (Ejm: "secure", "masterclass",
"prueba")}

\item{dias}{es el tiempo en días que se quiere ver hacia atrás. Por ejemplo
si
se quiere ver los tiquetes cerrados en los últimos 7 días, lag = 7. Por
defecto muestra todos los tiquetes sin ningún lag.}
}
\value{
No hay valor de retorno - se llama por su efecto secundario
}
\description{
Listado de tiquetes cerrados del repositorio indicado.
}
