% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_users.R
\name{new_user}
\alias{new_user}
\title{Create a new user on ixplorer.}
\usage{
new_user(base_url, api_key, owner, repo, user_data)
}
\arguments{
\item{base_url}{The base URL of the ixplorer server.}

\item{api_key}{The API key for authentication on the ixplorer server.}

\item{owner}{The owner (organization or user) of the repository where
the user will be created.}

\item{repo}{The name of the repository where the user will be created.}

\item{user_data}{An object containing user information, including
at least email, login_name, and username.}
}
\value{
A list containing the response from the ixplorer server,
including the generated password.
}
\description{
This function creates a new user on an ixplorer server by sending a POST
request to the appropriate endpoint.
}
