% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_wiki.R
\name{publica_wiki}
\alias{publica_wiki}
\title{Publica en el wiki repo}
\usage{
publica_wiki(
  reporte = report,
  ruta_repo_wiki = path_wiki_repo,
  auto_actualizar = TRUE,
  silenciar_bitacora = FALSE
)
}
\arguments{
\item{reporte}{la ruta al archivo rmarkdown o qmd}

\item{ruta_repo_wiki}{la ruta completa al repositorio wiki.}

\item{auto_actualizar}{Si desea hacer automáticamente el pull, commit y push
establecido en TRUE, si desea hacerlo manualmente, establezca FALSE. El
valor predeterminado es TRUE.}

\item{silenciar_bitacora}{Si desea ver la bitacora de renderización
establezca TRUE, de lo contrario establezca FALSE. El valor default es TRUE}
}
\value{
Publica un md en un repositorio wiki.
}
\description{
Publique un rmarkdown o qmd específico en el repositorio wiki,
especificando la ruta del repositorio wiki una vez que ya haya clonado este
repositorio en su máquina. Consulte Detalles si es la primera vez que
publica en un repositorio wiki.
}
\details{
Publica un rmarkdown en un wiki repo

Si es la primera vez que va a publicar en una wiki,  se deben seguir
estos pasos:

1.El repositorio wiki debe clonarse en una carpeta de trabajo local.
Para hacer esto, primero vaya a la pestaña wiki de su repositorio
y deberá crear manualmente la primera página wiki. Si no lo hace,
le dirá que el repositorio no existe.
\enumerate{
\item Una vez que haya creado la página de inicio de la wiki,
deberá clonar el repositorio de la wiki
como lo haría con cualquier otro repositorio en su computadora
(usando el enlace que está disponible para usted una vez que haya creado la
página de inicio de la wiki).
\item Una vez que haya clonado la wiki en su computadora, lo que debe indicar
en la función \code{publica_wiki} es la\code{ path_wiki_repo}. Esta es la dirección
completa en su computadora donde se encuentra este repositorio que clonó.
Por ejemplo:  \verb{/home/client/project/wiki_repo}
}
}
\examples{

\dontrun{
publish_wiki(reporte = "eda.Rmd",
             path_wiki_repo = "/home/cliente/proyecto/wiki_repo")
}

}
