% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{set_authentication}
\alias{set_authentication}
\title{Verifique cada uno de los pasos}
\usage{
set_authentication(access_data)
}
\description{
Verifique cada uno de los elementos necesarios para acceder a su repositorio
desde el gadget.
}
