% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_cache.R
\name{set_git_credentials_cache}
\alias{set_git_credentials_cache}
\title{Set git credential cache time}
\usage{
set_git_credentials_cache(hours = 4, global = TRUE)
}
\arguments{
\item{hours}{the amount of hours that the credentials should be stored.
Defaults to 4 hours.}

\item{global}{if TRUE then the setting will be sett as a global setting. If
FALSE it will only be for the current git repository}
}
\description{
Set the amount of time that the git credentials will be
available on the system you are working on.
}
\details{
As we move between servers to work we oftentimes do not want to
store our git passwords there. For instance when we are workign on servers
of third parties. The extreme is to include the username and password for
every single git call. But with git we can have a middle way where we store
the credentials in cache for a limited amount of time.
}
