% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_branch.R
\name{synch_branch}
\alias{synch_branch}
\title{Synch your master branch with upstream}
\usage{
synch_branch(default_branch_name = "master")
}
\value{

}
\description{
Runs steps to synch your master branch to upstream
automatically
}
