% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{verify_ixplorer_file}
\alias{verify_ixplorer_file}
\title{Verificar autentificación del ixplorer.}
\usage{
verify_ixplorer_file()
}
\description{
Verifique si hay un archivo .ixplorer en el directorio de trabajo y
configure las variables.
}
