% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{verify_ixproject}
\alias{verify_ixproject}
\title{Verify  ixproject}
\usage{
verify_ixproject(gitear_access)
}
\description{
Verify if there is a project name where your ixplorer repository belongs.
}
