% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{verify_ixrepo}
\alias{verify_ixrepo}
\title{Verifies ixplorer repo}
\usage{
verify_ixrepo(gitear_access)
}
\arguments{
\item{gitear_access}{parameters to access into an ixplorer instance}
}
\description{
Verifies if ixplorer repository's name exists.
}
