% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{verify_ixtoken}
\alias{verify_ixtoken}
\title{Verifies ixplorer user}
\usage{
verify_ixtoken(gitear_access)
}
\arguments{
\item{gitear_access}{parameters to access into an ixplorer instance}
}
\description{
Verifies token exists for a repository.
}
