% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{verify_ixurl}
\alias{verify_ixurl}
\title{Verifies ixplorer url}
\usage{
verify_ixurl(gitear_access)
}
\arguments{
\item{gitear_access}{parameters to access into an ixplorer instance}
}
\description{
Verifies if the url exists in the repository
}
