% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{verify_ixuser}
\alias{verify_ixuser}
\title{Verifies ixplorer user}
\usage{
verify_ixuser(gitear_access)
}
\arguments{
\item{gitear_access}{parameters to access into an ixplorer instance}
}
\description{
Verifies if user name exists.
}
