% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{add_token}
\alias{add_token}
\title{Autentificación en ixplorer}
\usage{
add_token()
}
\description{
Guarda en su computadora por medio del sistema de
 autentificación del OS de su computadora.
}
\details{
En caso de que los credenciales ya existan solo se ingresa el url
y se confirma si quiere conservar los credenciales en su computador o quiere
borrarlos luego de la siguiente consulta.
}
