% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{add_token}
\alias{add_token}
\title{Authenticate to ixplorer}
\usage{
add_token()
}
\description{
Make the connection to your repository through the ixplorer gadget and be
able to create tickets, review tickets without re-writing your credentials
}
