% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ticket.R
\name{create_tickets}
\alias{create_tickets}
\title{Crear tiquete}
\usage{
create_tickets(instance, owner, repository = "current")
}
\arguments{
\item{instance}{instancia de ixplorer (Ejm: "secure", "masterclass", "prueba")}

\item{owner}{el nombre del proyecto donde se encuentra el repositorio en
ixplorer}

\item{repository}{el nombre del repositorio donde están los tiquetes}
}
\description{
Crear tiquetes (título y cuerpo) desde el addin de ixplorer sin
perder las ideas durante su flujo de trabajo.
}
