% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ticket.R
\name{create_tickets}
\alias{create_tickets}
\title{Create ticket}
\usage{
create_tickets()
}
\description{
Create tickets (Title and body) from the ixplorer addin without loosing the
ideas during your workflow. tickets will be in the repository that corresponds
to the information youo give in the authenticate gadget.
}
