% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ticket.R
\name{create_tickets}
\alias{create_tickets}
\title{Create ticket}
\usage{
create_tickets(instance = "saved")
}
\arguments{
\item{instance}{ixplorer instance (Eg: "secure", "masterclass", "prueba")}

\item{owner}{the name of the project where the repository is located in ixplorer}

\item{repository}{the name of the repository where the tickets are}
}
\description{
Create tickets (title and body) from the ixplorer addin without
losing ideas during your workflow.
}
