% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_users.R
\name{create_users}
\alias{create_users}
\title{Create users and add them to an ixplorer repository.}
\usage{
create_users(
  base_url,
  api_key,
  owner,
  repo,
  user_data_df,
  collaborator_permissions
)
}
\arguments{
\item{base_url}{The base URL of the ixplorer server.}

\item{api_key}{The API key for authentication on the ixplorer server.}

\item{owner}{The owner (organization or user) of the repository where
the users will be added as collaborators.}

\item{repo}{The name of the repository where the users will be added
as collaborators.}

\item{user_data_df}{A data frame, each row containing user information
(email, login_name, username).}

\item{collaborator_permissions}{The permissions to be assigned
to the collaborators ("Read", "Write", or "Admin").}
}
\value{
A data frame containing information about the created users.
}
\description{
This function automates the process of creating users and adding
them as collaborators to a repository on ixplorer.
}
\examples{
\dontrun{
  # API configuration
  base_url <- "https://prueba.ixpantia.com"
  api_key <- "your_api_key" # Replace with your actual API key
  owner <- ""
  repo <- ""

   #User date in a data frame
   user_data_df <- data.frame(
   email = c("user@email.com", "user2@email.com"),
   login_name = c("User-1", "User-2"),
   username =  c("User-1", "User-2")
            )
  # Collaborator permissions to be added (Read, Write, or Admin)
  collaborator_permissions <- "Write"

  # Create users and add them to the repository with specific permissions
  results <- create_users(base_url, api_key, owner, repo, user_data_df,
  collaborator_permissions)}
}
