% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_tickets.R
\name{current_tickets}
\alias{current_tickets}
\title{Current tickets}
\usage{
current_tickets(instance, owner, repository = "current")
}
\arguments{
\item{instance}{ixplorer instance (Eg: "secure", "masterclass", "prueba")}

\item{owner}{the name of the project where the repository is located in ixplorer}

\item{repository}{the name of the repository where the tickets are}
}
\description{
View tickets for a specific user on a computer and get links to
your ixplorer based on the credentials used in the authentication gadget.
}
