% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_tickets.R
\name{current_tickets}
\alias{current_tickets}
\title{Tiquetes actuales}
\usage{
current_tickets()
}
\description{
Visualice los tiquetes de un usuario en específico, de un equipo y obtenga
los enlaces a su ixplorer basado en las credenciales utilizadas en el
gadget de autentificación.
}
