% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{incluye_upstream}
\alias{incluye_upstream}
\title{Title Agregue el upstream al repositorio que está activo en RStudio.}
\usage{
incluye_upstream(instancia = "guardada")
}
\arguments{
\item{instancia}{instancia de ixplorer (Ej: "secure", "masterclass",
"prueba")}
}
\value{
No hay valor de retorno - se llama por su efecto secundario
}
\description{
Title Agregue el upstream al repositorio que está activo en RStudio.
}
\details{
Esta función es la versión en español de add_upstream().
Perminte agregar el upstream al repositorio que está activo en RStudio en
este momento
}
