% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixplorer.R
\name{ix_issues}
\alias{ix_issues}
\title{ixplorer reports}
\usage{
ix_issues()
}
\description{
Visualize the issues of an specific user, a team and closed issues based on
the credentials used in gadget authenticate.
}
