% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{list_closed_tickets}
\alias{list_closed_tickets}
\title{list closed tickets}
\usage{
list_closed_tickets(lag = 7, repository = "current")
}
\description{
list tickets. the "current" respository is the one linked to the active
RStudio project.
}
