% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{list_closed_tickets}
\alias{list_closed_tickets}
\title{Enlistar tiquetes cerrados}
\usage{
list_closed_tickets(lag = 7, repository = "current", clip = TRUE)
}
\description{
Listado de los tiquetes. El repositorio actual es el que se encuentra
ligado al proyecto activo en RStudio.
}
