% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{list_open_tickets}
\alias{list_open_tickets}
\title{list open tickets}
\usage{
list_open_tickets(lag = 7, repository = "current", clip = TRUE)
}
\description{
list tickets. the "current" respository is the one linked to the active
RStudio project.
}
