% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_closed_tickets.R
\name{listar_tiquetes_cerrados}
\alias{listar_tiquetes_cerrados}
\title{Lista de tiquetes cerrados}
\usage{
listar_tiquetes_cerrados(
  instancia,
  propietario,
  repositorio = "actual",
  lag = NULL
)
}
\arguments{
\item{lag}{es el tiempo en días que se quiere ver hacia atrás. Por ejemplo si
se quiere ver los tiquetes cerrados en los últimos 7 días, lag = 7. Por
defecto muestra todos los tiquetes sin ningún lag.}

\item{instance}{instancia de ixplorer (Ejm: "secure", "masterclass", "prueba")}

\item{owner}{el nombre del proyecto donde se encuentra el repositorio en
ixplorer}

\item{repository}{el nombre del repositorio donde están los tiquetes}
}
\description{
Listado de tiquetes cerrados del repositorio indicado.
}
