% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{set_authentication}
\alias{set_authentication}
\title{Set authentication}
\usage{
set_authentication(access_data)
}
\arguments{
\item{access_data}{Data needed to get into ixplorer}
}
\value{
No return value, called for side effects
}
\description{
Set the authentification into ixplorer
}
