% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_tickets.R
\name{tiquetes_actuales}
\alias{tiquetes_actuales}
\title{Tiquetes actuales}
\usage{
tiquetes_actuales(instancia = "guardada")
}
\arguments{
\item{instancia}{instancia de ixplorer (Ej: "secure", "masterclass",
"prueba")}
}
\value{
No hay valor de retorno - se llama por su efecto secundario
}
\description{
Vea tiquetes para un usuario específico en una computadora y
obtenga enlaces a su ixplorer según las credenciales utilizadas en el gadget
de autenticación.
}
