% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{verify_ixplorer_file}
\alias{verify_ixplorer_file}
\title{Verify authentication to ixplorer}
\usage{
verify_ixplorer_file()
}
\description{
Verify if there is  a .ixplorer file in your working directory and set
the variables.
}
