% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{verify_ixproject}
\alias{verify_ixproject}
\title{Verifies ixplorer project}
\usage{
verify_ixproject(gitear_access)
}
\arguments{
\item{gitear_access}{parameters to access into an ixplorer instance}
}
\description{
Verifies if the project's name exists in the repository.
}
