% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_authenticacion.R
\name{verify_ixproject}
\alias{verify_ixproject}
\title{Verifique  ixproject}
\usage{
verify_ixproject(gitear_access)
}
\description{
Verifique si existe el nombre del proyecto donde pertenece el repositorio.
}
